+++ src/Exception/ErrorResponseException.php
@@ -6,14 +6,20 @@ namespace ErickSkrauch\Fcm\Exception;
 use Exception;
 use Psr\Http\Message\ResponseInterface;
 
-final class UnexpectedResponseException extends Exception {
+final class ErrorResponseException extends Exception {
+
+    /**
+     * @readonly
+     */
+    public string $errorCode;
 
     /**
      * @readonly
      */
     public ResponseInterface $response;
 
-    public function __construct(string $message, ResponseInterface $response) {
+    public function __construct(string $message, string $errorCode, ResponseInterface $response) {
+        $this->errorCode = $errorCode;
         $this->response = $response;
         parent::__construct($message, $response->getStatusCode());
     }