+++ tests/Recipient/DeviceTest.php
@@ -4,7 +4,6 @@ declare(strict_types=1);
 namespace ErickSkrauch\Fcm\Tests\Recipient;
 
 use ErickSkrauch\Fcm\Recipient\Device;
-use ErickSkrauch\Fcm\Recipient\Recipient;
 use PHPUnit\Framework\TestCase;
 
 /**
@@ -14,7 +13,7 @@ final class DeviceTest extends TestCase {
 
     public function test(): void {
         $model = new Device('mock token');
-        $this->assertSame(Recipient::PARAM_TO, $model->getConditionParam());
+        $this->assertSame('token', $model->getConditionParam());
         $this->assertSame('mock token', $model->getConditionValue());
     }